videoPlayer controls custom

+++ controllers/videoController.js
@@ -65,7 +65,11 @@ export const getEditVideo = async (req, res) => {
   } = req;
   try {
     const video = await Video.findById(id);
-    res.render("editVideo", { pageTitle: `Edit ${video.title}`, video });
+    if (video.creator != req.user.id) {
+      throw Error();
+    } else {
+      res.render("editVideo", { pageTitle: `Edit ${video.title}`, video });
+    }
   } catch (error) {
     res.redirect(routes.home);
   }
@@ -89,12 +93,17 @@ export const deleteVideo = async (req, res) => {
     params: { id }
   } = req;
   try {
-    //  upload/videos data remove code
-    const fileUrl = await Video.find({ _id: id }, "fileUrl");
-    const fileUrlPath = fileUrl[0].fileUrl;
-    fs.unlinkSync(fileUrlPath);
-    //  mongodb data remove code
-    await Video.findOneAndRemove({ _id: id });
+    const video = await Video.findById(id);
+    if (video.creator != req.user.id) {
+      throw Error();
+    } else {
+      //  upload/videos data remove code
+      const fileUrl = await Video.find({ _id: id }, "fileUrl");
+      const fileUrlPath = fileUrl[0].fileUrl;
+      fs.unlinkSync(fileUrlPath);
+      //  mongodb data remove code
+      await Video.findOneAndRemove({ _id: id });
+    }
   } catch (error) {
     console.log(error);
   }
